Split osm_store, introduce db-storage feature
The file `src/osm_reader/street.rs` contained a large chunk of code
related to storage that was not related to streets per se. This part of
the code was moved to a separate file, `src/osm_reader/osm_store.rs`.

In addition, the dependency on a database storage (Sqlite) was made
optional with the help of a feature 'db-storage'.
