+++ libs/mimir/Cargo.toml
@@ -16,9 +16,6 @@ rs-es = { git = "https://github.com/canaltp/rs-es", features = ["geo"]}
 serde = { version = "1", features = ["rc"]}
 serde_json = "1"
 chrono = "0.4"
-# we are stuck with reqwest 0.9.16 because later version does not work in an async context (such as actix)
-# we would need to either make rs-es async or remove rs-es to update reqwest
-# cf https://github.com/benashford/rs-es/issues/136
 reqwest = { version = "0.10", features = [ "blocking", "json" ] }
 geo = "0.14"
 geo-types = { version = "0.6", features = [ "rstar" ] }
