+++ src/osm_reader/mod.rs
@@ -3,6 +3,7 @@ use std::fs::File;
 use std::path::Path;
 
 pub mod admin;
+pub mod osm_store;
 pub mod osm_utils;
 pub mod poi;
 pub mod street;
